Create a sandbox for user's code.
* Put an iframe element in the app componet to create a sandbox for users code.
This will make the code safer.  Now if the user executes erroneous code, or
if he tries to manipulate DOM, or if he tries to execute some malicious code,
then our code will be safe (because of different context and sandboxed code).

* Use the srcDoc props in the iframe and set it to a pre-written html string.
This will make the document sandboxed and it will lack the 'allow-same-origin'
falg, hence some features like localStorage won't be available.